requirements.txt
python-dotenv
pandas
openpyxl
paddlepaddle
paddleocr
opencv-python
google
requests
matplotlib
jupyter
notebook
torch 
torchvision 
transformers 
pillow 
sentencepiece